Stand-alone apply
