+++ doc/code/qp_core_apply.rst
@@ -1,0 +1,9 @@
+qp.core.apply
+=============
+
+Overview
+--------
+
+.. currentmodule:: pennylane.core.apply
+
+.. automodule:: pennylane.core.apply
