+++ doc/index.rst
@@ -273,6 +273,7 @@ PennyLane is **free** and **open source**, released under the Apache License, Ve
 
    code/qp_capture
    code/qp_concurrency
+   code/qp_core_apply
    code/qp_core_operator
    code/qp_core_queuing
    code/qp_devices
