+++ pennylane/core/operator/operator2.py
@@ -33,7 +33,8 @@ import pennylane as qp
 from pennylane import math
 from pennylane._class_property import classproperty
 from pennylane.capture import enabled, pause, symbolic_array
-from pennylane.core.queuing import AnnotatedQueue, QueuingManager, apply
+from pennylane.core.apply import apply
+from pennylane.core.queuing import AnnotatedQueue, QueuingManager
 from pennylane.exceptions import (
     AdjointUndefinedError,
     DecompositionUndefinedError,
