+++ pennylane/core/queuing.py
@@ -26,7 +26,6 @@ This module contains the classes for placing objects into queues.
 
     ~QueuingManager
     ~AnnotatedQueue
-    ~apply
 
 
 Description
@@ -129,8 +128,8 @@ not do anything.
 >>> q.queue
 [X(0)]
 
-The :func:`~.apply` method allows a single object to be queued multiple times in a circuit.
-The function queues a copy of the original object if it already in the queue.
+The :func:`~pennylane.apply` function allows a single object to be queued multiple times in a
+circuit. The function queues a copy of the original object if it is already in the queue.
 
 >>> op = qp.X(0)
 
@@ -188,13 +187,12 @@ use the :meth:`~.queuing.QueuingManager.stop_recording` context upon constructio
 
 """
 
-import copy
 from collections import OrderedDict
 from contextlib import contextmanager
 from threading import RLock
 from typing import Optional
 
-from pennylane import capture, pytrees  # tach-ignore
+from pennylane import capture  # tach-ignore
 from pennylane.exceptions import QueuingError
 
 
@@ -442,134 +440,6 @@ class AnnotatedQueue(OrderedDict):
         return super().__contains__(key)
 
 
-def apply(op, context: type[QueuingManager] | AnnotatedQueue = QueuingManager):
-    """Apply an instantiated operator or measurement to a queuing context.
-
-    Args:
-        op (.Operator or .MeasurementProcess): the operator or measurement to apply/queue
-        context (type[.QueuingManager] | AnnotatedQueue): The queuing context to queue the operator to.
-            Note that if no context is specified, the operator is
-            applied to the currently active queuing context.
-    Returns:
-        .Operator or .MeasurementProcess: the input operator is returned for convenience
-    Raises:
-        RuntimeError: if we try to use apply in a non-queuing/non-tracing context.
-
-    **Example**
-
-    In PennyLane, operations and measurements are 'queued' or added to a circuit
-    when they are instantiated.
-
-    The ``apply`` function can be used to add operations that might have
-    already been instantiated elsewhere to the QNode:
-
-    .. code-block:: python
-
-        op = qp.RX(0.4, wires=0)
-        dev = qp.device("default.qubit", wires=2)
-
-        @qp.qnode(dev)
-        def circuit(x):
-            qp.RY(x, wires=0)  # applied during instantiation
-            qp.apply(op)  # manually applied
-            return qp.expval(qp.Z(0))
-
-    >>> print(qp.draw(circuit)(0.6))
-    0: ──RY(0.60)──RX(0.40)─┤  <Z>
-
-    It can also be used to apply functions repeatedly:
-
-    .. code-block:: python
-
-        @qp.qnode(dev)
-        def circuit(x):
-            qp.apply(op)
-            qp.RY(x, wires=0)
-            qp.apply(op)
-            return qp.expval(qp.Z(0))
-
-    >>> print(qp.draw(circuit)(0.6))
-    0: ──RX(0.40)──RY(0.60)──RX(0.40)─┤  <Z>
-
-    .. warning::
-
-        If you use ``apply`` on an operator that has already been queued, it will
-        be queued for a second time. For example:
-
-        .. code-block:: python
-
-            @qp.qnode(dev)
-            def circuit():
-                op = qp.Hadamard(0)
-                qp.apply(op)
-                return qp.expval(qp.Z(0))
-
-        >>> print(qp.draw(circuit)())
-        0: ──H──H─┤  <Z>
-
-    .. details::
-        :title: Usage Details
-
-        Instantiated measurements can also be applied to queuing contexts
-        using ``apply``:
-
-        .. code-block:: python
-
-            meas = qp.expval(qp.Z(0) @ qp.Y(1))
-            dev = qp.device("default.qubit", wires=2)
-
-            @qp.qnode(dev)
-            def circuit(x):
-                qp.RY(x, wires=0)
-                qp.CNOT(wires=[0, 1])
-                return qp.apply(meas)
-
-        >>> print(qp.draw(circuit)(0.6))
-        0: ──RY(0.60)─╭●─┤ ╭<Z@Y>
-        1: ───────────╰X─┤ ╰<Z@Y>
-
-        By default, ``apply`` will queue operators to the currently
-        active queuing context.
-
-    """
-
-    if capture.enabled():
-        return _capture_apply(op)
-
-    if not QueuingManager.recording():
-        raise RuntimeError("No queuing context available to append operation to.")
-
-    # Always make a copy since we don't want the provided op to be dequeued by a subsequent
-    # PennyLane Operator/Function.
-    # Note that queuing contexts can only contain unique objects.
-    with QueuingManager.stop_recording():
-        op = copy.copy(op)
-
-    if hasattr(op, "queue"):
-        # operator provides its own logic for queuing
-        op.queue(context=context)
-    else:
-        # append the operator directly to the relevant queuing context
-        context.append(op)
-
-    return op
-
-
-def _capture_apply(op):
-    """Applies an op in a capture context."""
-
-    if hasattr(op, "_bind_primitive"):
-        op._bind_primitive()  # pylint: disable=protected-access
-        if op.tracer is None:
-            raise RuntimeError("Trying to use apply in a non-tracing context.")
-        return op
-
-    # Capture is active but the op has no _bind_primitive (e.g. minimal
-    # legacy Operator subclass).  Reconstruct via the constructor so the
-    # new instance auto-binds its primitive.
-    return pytrees.unflatten(*pytrees.flatten(op))
-
-
 def remove_from_program(op):
     """Removes an operator from the captured/queued program."""
     if QueuingManager.recording():
@@ -581,4 +451,4 @@ def remove_from_program(op):
         pop_op_eqns((op,))
 
 
-__all__ = ["QueuingManager", "AnnotatedQueue", "apply", "remove_from_program"]
+__all__ = ["QueuingManager", "AnnotatedQueue", "remove_from_program"]
