+++ pennylane/labs/templates/left_classical_comparator.py
@@ -15,8 +15,9 @@
 test of a quantum register and a classical integer."""
 
 from pennylane import capture, compiler, cond, for_loop, math
+from pennylane.core.apply import apply
 from pennylane.core.operator import Operation
-from pennylane.core.queuing import AnnotatedQueue, QueuingManager, apply
+from pennylane.core.queuing import AnnotatedQueue, QueuingManager
 from pennylane.decomposition import (
     add_decomps,
     register_resources,
