+++ pennylane/labs/templates/left_quantum_comparator.py
@@ -14,8 +14,9 @@
 """Contains the LeftQuantumComparator template for performing inequality test of two quantum registers."""
 
 from pennylane import capture, compiler, for_loop, math
+from pennylane.core.apply import apply
 from pennylane.core.operator import Operation
-from pennylane.core.queuing import AnnotatedQueue, QueuingManager, apply
+from pennylane.core.queuing import AnnotatedQueue, QueuingManager
 from pennylane.decomposition import (
     add_decomps,
     register_resources,
