+++ pennylane/ops/op_math/change_op_basis.py
@@ -23,6 +23,7 @@ from functools import reduce
 
 from pennylane import capture, math
 from pennylane.core import queuing
+from pennylane.core.apply import apply
 from pennylane.core.operator import Operator, Operator2, abstractify
 from pennylane.core.operator.operator2 import pop_op_eqns  # tach-ignore
 from pennylane.decomposition import add_decomps, register_resources
@@ -80,7 +81,7 @@ def _apply_op_or_func(op_or_func):
             # pylint: disable-next=protected-access
             op_or_func._bind_primitive()
     elif isinstance(op_or_func, Operator):
-        queuing.apply(op_or_func)
+        apply(op_or_func)
     elif _is_abstract_operator(op_or_func):
         pass
     else:
@@ -364,7 +365,7 @@ class ChangeOpBasis(CompositeOp):
     def decomposition(self):
         r"""Decomposition of the product operator is given by each of compute_op, target_op, compute_op† applied in succession."""
         if queuing.QueuingManager.recording():
-            _ = [queuing.apply(op) for op in reversed(self)]
+            _ = [apply(op) for op in reversed(self)]
         return list(self[::-1])
 
     # pylint: disable=arguments-renamed, invalid-overridden-method
@@ -404,9 +405,9 @@ def _adjoint_change_op_basis_resources(base_params, **_):
 # pylint: disable=protected-access
 @register_resources(_adjoint_change_op_basis_resources)
 def _adjoint_change_op_basis_decomp(*_, base, **__):
-    queuing.apply(base.operands[2])
-    adjoint(queuing.apply(base.operands[1]))
-    queuing.apply(base.operands[0])
+    apply(base.operands[2])
+    adjoint(apply(base.operands[1]))
+    apply(base.operands[0])
 
 
 add_decomps("Adjoint(ChangeOpBasis)", _adjoint_change_op_basis_decomp)
@@ -447,22 +448,22 @@ def _controlled_change_op_basis_decomposition(
     base,
     **__,
 ):
-    queuing.apply(base.operands[2])
+    apply(base.operands[2])
     ctrl(
-        queuing.apply(base.operands[1]),
+        apply(base.operands[1]),
         control=control_wires,
         control_values=control_values,
         work_wires=work_wires,
         work_wire_type=work_wire_type,
     )
-    queuing.apply(base.operands[0])
+    apply(base.operands[0])
 
 
 # pylint: disable=unused-argument
 @register_resources(_change_op_basis_resources)
 def _change_op_basis_decomp(*_, wires=None, operands, **__):
     for op in operands[::-1]:
-        queuing.apply(op)
+        apply(op)
 
 
 add_decomps(ChangeOpBasis, _change_op_basis_decomp)
