+++ pennylane/ops/op_math/exp.py
@@ -24,6 +24,7 @@ from scipy.sparse.linalg import expm as sparse_expm
 import pennylane as qp
 from pennylane import math
 from pennylane.core import queuing
+from pennylane.core.apply import apply
 from pennylane.core.operator import Operation, Operator
 from pennylane.decomposition import (
     add_decomps,
@@ -311,7 +312,7 @@ class Exp(ScalarSymbolicOp, Operation):
 
         if queuing.QueuingManager.recording():
             for op in d:
-                queuing.apply(op)
+                apply(op)
 
         return d
