+++ pennylane/optimize/shot_adaptive.py
@@ -23,8 +23,8 @@ import scipy as sp
 
 from pennylane import math, measurements
 from pennylane._grad import jacobian
+from pennylane.core.apply import apply
 from pennylane.core.qscript import make_qscript
-from pennylane.core.queuing import apply
 from pennylane.ops import LinearCombination
 from pennylane.workflow import QNode, construct_tape, set_shots
