+++ pennylane/templates/subroutines/select.py
@@ -22,8 +22,9 @@ import numpy as np
 
 from pennylane import math
 from pennylane import ops as qp_ops
+from pennylane.core.apply import apply
 from pennylane.core.operator import Operation, Operator2, abstractify
-from pennylane.core.queuing import QueuingManager, apply
+from pennylane.core.queuing import QueuingManager
 from pennylane.decomposition import add_decomps, register_condition, register_resources
 from pennylane.ops import CNOT, X, adjoint, ctrl
 from pennylane.ops.op_math.adjoint2 import _adjoint_abstract
