+++ pennylane/templates/subroutines/time_evolution/approx_time_evolution.py
@@ -20,8 +20,9 @@ import copy
 from collections import defaultdict
 
 from pennylane.control_flow import for_loop
+from pennylane.core.apply import apply
 from pennylane.core.operator import Operation
-from pennylane.core.queuing import QueuingManager, apply
+from pennylane.core.queuing import QueuingManager
 from pennylane.decomposition import add_decomps, register_resources
 from pennylane.ops import PauliRot
 from pennylane.ops.op_math.linear_combination import Hamiltonian
