+++ pennylane/templates/subroutines/time_evolution/commuting_evolution.py
@@ -19,8 +19,9 @@ Contains the CommutingEvolution template.
 import copy
 
 from pennylane import capture, math
+from pennylane.core.apply import apply
 from pennylane.core.operator import Operation, Operator1
-from pennylane.core.queuing import QueuingManager, apply
+from pennylane.core.queuing import QueuingManager
 from pennylane.decomposition import add_decomps, register_resources, resource_rep
 from pennylane.ops.op_math.linear_combination import Hamiltonian
 from pennylane.pauli import PauliWord
