+++ pennylane/templates/subroutines/time_evolution/trotter.py
@@ -19,9 +19,10 @@ import copy
 
 from pennylane import ops as qp_ops
 from pennylane.capture.autograph import wraps
+from pennylane.core.apply import apply
 from pennylane.core.operator import Operation, Operator
 from pennylane.core.qscript import make_qscript
-from pennylane.core.queuing import QueuingManager, apply
+from pennylane.core.queuing import QueuingManager
 from pennylane.decomposition import add_decomps, register_resources, resource_rep
 from pennylane.wires import Wires
