+++ pennylane/templates/tensornetworks/mera.py
@@ -23,9 +23,10 @@ from collections.abc import Callable
 import numpy as np
 
 from pennylane import math
+from pennylane.core.apply import apply
 from pennylane.core.operator import Operation
 from pennylane.core.qscript import make_qscript
-from pennylane.core.queuing import QueuingManager, apply
+from pennylane.core.queuing import QueuingManager
 
 
 def compute_indices(wires, n_block_wires):
