+++ pennylane/templates/tensornetworks/ttn.py
@@ -22,9 +22,10 @@ import warnings
 import numpy as np
 
 from pennylane import math
+from pennylane.core.apply import apply
 from pennylane.core.operator import Operation
 from pennylane.core.qscript import make_qscript
-from pennylane.core.queuing import QueuingManager, apply
+from pennylane.core.queuing import QueuingManager
 
 
 def compute_indices(wires, n_block_wires):
